Fix missing dist files based on .tsx

+++ scripts/clean-dist.sh
@@ -13,10 +13,13 @@ for f in `find dist -type f | grep \\\.mjs$`
 do
   src_f="_src/${f#dist/}"
   src_f="${src_f%.mjs}.mts"
+  src_f2="${src_f}x"
   raw="${f%.mjs}"
   if [ ! -f "$src_f" ]; then
-    echo "Removing $raw.mjs"
-    rm -f $raw.mjs $raw.mjs.map $raw.d.mts $raw.d.mts.map
+    if [ ! -f "$src_f2" ]; then
+      echo "Removing $raw.mjs"
+      rm -f $raw.mjs $raw.mjs.map $raw.d.mts $raw.d.mts.map
+    fi
   fi
 done
 
@@ -25,9 +28,12 @@ for f in `find dist -type f | grep \\\.js$`
 do
   src_f="_src/${f#dist/}"
   src_f="${src_f%.js}.ts"
+  src_f2="${src_f}x"
   raw="${f%.js}"
   if [ ! -f "$src_f" ]; then
-    echo "Removing $raw.js"
-    rm -f $raw.js $raw.js.map $raw.d.ts $raw.d.ts.map
+    if [ ! -f "$src_f2" ]; then
+      echo "Removing $raw.js"
+      rm -f $raw.js $raw.js.map $raw.d.ts $raw.d.ts.map
+    fi
   fi
 done
